easier to use

+++ main.py
@@ -11,6 +11,9 @@ import time
 
 # Real gen se pehle timer on kar dena pls
 
+data_location = ""
+cert_location = ""
+mail = True
 timer = False
 print("Started")
 sender = 'jainnimansh@gmail.com'
@@ -33,16 +36,16 @@ def sendMail(receiver1, sample_pdf_path, name):
     <html>
         <body>
             <h1>Hi {name},</h1>
-            Thank you for participating in <a href="https://techsyndicate.us/intech" target="_blank">inTech '22</a> - the 4th edition of Tech Syndicate's annual intra school tech event.
-            This year we had more than 200+ registrations and received 150+ submissions. We hope you enjoyed participating in this competition as much as we enjoyed organising it. 
+            Thank you for participating in <a href="https://techsyndicate.us/robo" target="_blank">Robotronics '22</a> - the 10th edition of Tech Syndicate's annual tech event.
+            This year we had more than 500+ registrations and received 150+ submissions. We hope you enjoyed participating in this competition as much as we enjoyed organising it. 
             <br><br>
-            Congratulations and best of luck for the future. Stay tuned for inTech '23.
+            Congratulations and best of luck for the future. Stay tuned for Robotronics '23.
             <br><br>
             Please find your certificates attached and feel free to DM any moderator in case of issues. 
             <br><br>
             Revolutionize<br>
-            Aayan Agarwal, President <a href="tel:+91 9650573555">(+91 9650573555)</a><br>
-            Aayush Garg, Vice President <a href="tel:+91 9717966964">(+91 9717966964)</a><br>
+            Aayan Agarwal, President (<a href="tel:+91 9650573555">+91 9650573555</a>)<br>
+            Aayush Garg, Vice President (<a href="tel:+91 9717966964">+91 9717966964</a>)<br>
         </body>
     </html>  
     """.format(name=name)
@@ -67,18 +70,18 @@ def sendMail(receiver1, sample_pdf_path, name):
 
 def makePart(per):
     img = Image.open(
-        "cert 6.png", mode='r')
+        cert_location, mode='r')
     draw = ImageDraw.Draw(img)
 
     name = per["name"]
     font = ImageFont.truetype(
         "font.woff2",
-        100
+        80
     )
     txtSize = font.getlength(name)
     nameX = (4960-txtSize)/2
     draw.text(
-        (nameX, 982),
+        (nameX, 1001),
         name,
         fill='#1b1b1b',
         font=font)
@@ -86,11 +89,11 @@ def makePart(per):
     schooltxtsize = font.getlength(per["school"])
     font = ImageFont.truetype(
         "font.woff2",
-        100
+        80
     )
     schoolX = (5100-schooltxtsize)/2
     draw.text(
-        (schoolX, 1149),
+        (schoolX, 1170),
         per["school"],
         fill="#1b1b1b",
         font=font
@@ -99,11 +102,11 @@ def makePart(per):
     eventtxtsize = font.getlength(per["field"])
     font = ImageFont.truetype(
         "font.woff2",
-        100
+        80
     )
     EventX = (5650-eventtxtsize)/2
     draw.text(
-        (EventX, 1304),
+        (EventX, 1323),
         per["field"],
         fill="#1b1b1b",
         font=font
@@ -121,7 +124,7 @@ def conversion(flag):
 
 
 def generator():
-    f = open('data.json')
+    f = open(data_location)
     res = json.load(f)
     dt = res["data"]
 
@@ -133,19 +136,19 @@ def generator():
             print(e)
             print("Error making Participation Certificate for:", i["name"])
         end = time.time()
-        print("gen time", end-begin)
-        
-        begin = time.time()
-        try:
-            sendMail(i["mail"], "Partake.png", i["name"])
-        except Exception as e:
-            print(e)
-            print("Error Sending Mail to:", i["name"])
-        end = time.time()
-        print("send mail time", end-begin)
-        os.remove("Partake.png")
+        print("gen time", int(end-begin))
+        if mail:
+            begin = time.time()
+            try:
+                sendMail(i["mail"], "Partake.png", i["name"])
+            except Exception as e:
+                print(e)
+                print("Error Sending Mail to:", i["name"])
+            end = time.time()
+            print("send mail time", int(end-begin))
+            os.remove("Partake.png")
         if timer:
-            time.sleep(60)
+            time.sleep(30)
     f.close()
     session.quit()
     print("Finished")
